HOTFIX

+++ codeine_django/organization/views_contribution_payment.py
@@ -30,7 +30,7 @@ def contribution_payment_view(request):
 
         with transaction.atomic():
             try:
-                month_duration = int(data['month_duration'])
+                month_duration = float(data['month_duration'])
                 payment_transaction = PaymentTransaction(
                     payment_amount=float(data['contribution']) * month_duration,
                     payment_type=data['payment_type']
